requirements.txt
inferactively-pymdp==0.0.7.1
numpy>=1.23
matplotlib>=3.6
scipy>=1.9
